Some Esri methods require a referer
A quirk of some Esri API's is that they require a referer to be set.
The referer must be set in both the request for the token and then also
in the later API calls.

Unfortunately, it requires the referer to be set by different
mechanisms for these types of calls. The request for token requires a
querystring param, while the other API calls require it to be set in
the header.

The value doesn't really matter, so I'm setting the value to
`http://www.arcgis.com` in effort to get this out of the way as much as
possible.

It can be overridden with the `with-referrer` macro, though this
shouldn't be necessary in practice.